[api] Create a flask API for trackit
Trackit will be an internet reachable service for posting and collecting life
data. To achieve that end this change introduces the trackit API and a unit
test framework. The API is a simple flask API that can take in PUT and GET
requests for creating schemata and data and subsequently receiving them.