[proxy] Add a reverse proxy for the API
The flask server by default is not great and handling lots of requests but for
the time being we shouldn't really care as our load won't be very high. We will
be putting the API behind a load-balancing service that may keep multiple
connections open so this change adds an nginx layer between the internet and
the flask API so that we can more easily scale up our API containers and handle
keep-alive connections from the load balancer.