[trackit] e2e orchestration and testing
docker-compose is a nice tool for managing the multiple microservices that make
up your service. With this change we'll be able to use a single command to
build all of trackit, a single command to push up the distributed application
to a repository, and a single command to run it either locally or remotely.
Moreover this change adds e2e testing so it is possible to spin a trackit
instance locally and test integration between the services. There are also
quality tests to ensure adherence to formatting and linting standards.